+++ GTK_module/Makefile
@@ -1,21 +1,8 @@
-main : main.o gtk_module.o
-	g++ -o $@.out main.o gtk_module.o `pkg-config --cflags --libs gtk+-3.0` 
+main: main.o
+	g++ -o $@.out main.o `pkg-config gtkmm-3.0 --cflags --libs`
 
-main.o : main.cpp gtk_module.cpp gtk_module.hpp
-	g++ -c $*.cpp `pkg-config --cflags --libs gtk+-3.0`
-
-output.o : gtk_module.cpp gtk_module.hpp
-	g++ -c $*.cpp `pkg-config --cflags --libs gtk+-3.0`
+main.o: main.cpp
+	g++ -c $*.cpp `pkg-config gtkmm-3.0 --cflags --libs`
 
 clean:
-	rm -rf main.out main.o gtk_module.o
-
-#make:
-	#g++ output.cpp -o output.out `pkg-config --cflags --libs gtk+-3.0`
-	#g++ main.cpp -o main.out `pkg-config --cflags --libs gtk+-3.0`
-	#g++ output.backup.cpp -o output.backup.out `pkg-config --cflags --libs gtk+-3.0`
-#clean:
-	#rm -rf output.out
-
-#make :
-	#g++ -o main.out main.cpp output.cpp `pkg-config --cflags --libs gtk+-3.0`
+	rm -rf *.out *.o
