+++ GTK_module/legacy/main.cpp
@@ -1,0 +1,9 @@
+#include "gtk_module.hpp"
+
+int main(){
+    GTK_MODULE gtk;
+
+    gtk.init(NULL, NULL);
+
+    return 0;
+}
